LawFlow/requirements.txt
# Flask backend
flask==3.1.3
flask-cors==6.0.0

# Database
sqlalchemy==2.0.36

# Anthropic Claude API
anthropic==0.42.0

# Document processing
pdfplumber==0.11.4
python-pptx==1.0.2
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
uuid6==2024.7.10
